+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Added
-- `MergeMaps`, a `Monoid` on `Map` formed by `Map#merge` 
+- `MergeMaps`, a `Monoid` on `Map` formed by `Map#merge`
+- `CheckedEffect` is now a `CheckedFn1`
+- `CheckedSupplier` is now a `CheckedFn1`
+- `CheckedFn1` now overrides all possible methods with covariant return type 
 
 ### Fixed
 - issue where certain ways to compose `Effect`s unintentionally nullified the effect